+++ Modules/Inventario/Http/Controllers/MantenimientosProgramadosController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Modules\Inventario\Http\Controllers;
+
+use Illuminate\Routing\Controller;
+
+class MantenimientosProgramadosController extends Controller
+{
+    public function index()
+    {
+        return view('inventario::mantenimientos.index');
+    }
+}